chore(param editor): Add tests
